requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.6.0

# Machine Learning dependencies
lpips>=0.1.3
pytorch-msssim>=0.2.1
scikit-image>=0.18.0

# Utility dependencies
configparser
warnings
